New Fix for Frame refresh bug: http://groups.google.com/group/firebug/browse_thread/thread/c69e49b9815b2540 Issue 239
This page changes the frame href in a mouseover handler.  
This sends a mutate event for the document in the frame, but this document has no parentNode 
and yet it is not the rootElement for the HTML view.  So for nodes with no parentNodes and
nodeType 9 (document), return contentDocument.documentElement, the <html> element.

git-svn-id: http://fbug.googlecode.com/svn@69 e969d3be-0e28-0410-a27f-dd5c76401a8b