Clean up how context.sourceFileMap is built and used.
Back on 1.05 FBL.updateScriptFiles built and cache a list of source files. In 1.1 event and eval sources were added.  When looking at Chromebug to add chrome and XPCOM sources, the design of the updateScriptFiles had holes.  In 1.1, Multiple FBL.SourceFiles could be created for a single URL.  To fix this:
1) FBL.SourceFiles now takes a context and it sets 'this' into the context.sourceFileMap directly (one to one)
2) context.sourcefiles now is only a cache for FBL.updateScriptFiles and should not be used elsewhere.  
3) updateScriptFiles much simpler, it only deals with <script> tags. (?why does it even to that?)
4) eval and event analysis fills context.sourceFileMap

One side effect is that top level scripts come in from global/context/bindings/scrollbar.xml.  They seem to be legitimate scripts in the window. I will suppress these in the UI; I hope they are not a sign a bug.  The new Script panel option showAllSourceFiles was added to unsuppress the files.

git-svn-id: http://fbug.googlecode.com/svn@78 e969d3be-0e28-0410-a27f-dd5c76401a8b