trace to panel bug noThrottle not defined. &nbsp; showing under spinner on net tab, I changed it to a _.
Here's full story in case someone figures it out in the future. In
Firebug net.js the function formatTime(-1) returned "&nbsp;". This
value is inserted into a span with a class netTimeLabel; the
background image for that class is the spinner.  So the span must have
some content or the spinner will no show.  Sometime in the past the
value of &nbsp; worked. My guess is that some code was added to escape
the contents of the span (or every span in the table). So now the
actual HTML seen by FF not the entity but its literal.  So now we have
a funky string with a spinner. 



git-svn-id: http://fbug.googlecode.com/svn@121 e969d3be-0e28-0410-a27f-dd5c76401a8b