[&1.7] Issue 3704: 	Watch panel locals not listed. Exactly the same fix will work on 1.7, but the code on 1.7 is in lib.js.
Mozilla has a platform bug and we get a bogus scope. We try to operate on it and we get a powerful exception that blows off the JS stack.

The fix/workaround is here to 1) not touch the bogus object, put a message in its place. Should be low risk for 1.6.1

git-svn-id: http://fbug.googlecode.com/svn@8609 e969d3be-0e28-0410-a27f-dd5c76401a8b